requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
distlib==0.3.4
Django==3.2.10
django-crispy-forms==1.14.0
django-restframework==0.0.1
djangorestframework==3.13.1
filelock==3.7.1
fonttools==4.33.3
gunicorn==20.1.0
idna==3.3
kiwisolver==1.4.3
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
platformdirs==2.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
whitenoise==6.2.0
